requirements.txt
numpy
pandas
torch
matplotlib
jupyter
chardet
fire
patool
tqdm